backend/requirements_render.txt
# Django and Core Dependencies
Django==5.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Production Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# HTTP Requests
requests==2.31.0

# Security
cryptography==41.0.7

# AI Services - REQUIRED for full functionality
openai==1.3.7
google-generativeai==0.3.2
Pillow==10.1.0

# Cloud Storage - REQUIRED for file uploads
boto3==1.34.0
django-storages==1.14.2

# Payment Processing - REQUIRED for billing
stripe==7.8.0

# Task Queue - REQUIRED for background processing
celery==5.3.4
redis==5.0.1 